--- src/basic/App.tsx
@@ -1,59 +1,0 @@
-import React, { useState } from 'react';
-import { Header } from './components/common/layout/Header';
-import { AdminPage, CartPage } from './components';
-import { Coupon, Product } from '../types.ts';
-import { initialCoupons, initialProducts } from './mocks/data.ts';
-
-export const App: React.FC = () => {
-  const [products, setProducts] = useState<Product[]>(initialProducts);
-  const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
-  const [isAdmin, setIsAdmin] = useState(false);
-
-  //action
-  const handleProductUpdate = (updatedProduct: Product) => {
-    setProducts(prevProducts =>
-      prevProducts.map(p => p.id === updatedProduct.id ? updatedProduct : p)
-    );
-  };
-
-  //action
-  const handleProductAdd = (newProduct: Product) => {
-    setProducts(prevProducts => [...prevProducts, newProduct]);
-  };
-
-  //action
-  const handleCouponAdd = (newCoupon: Coupon) => {
-    setCoupons(prevCoupons => [...prevCoupons, newCoupon]);
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-100">
-      <nav className="bg-blue-600 text-white p-4">
-        <div className="container mx-auto flex justify-between items-center">
-          <h1 className="text-2xl font-bold">쇼핑몰 관리 시스템</h1>
-          <button
-            onClick={() => setIsAdmin(!isAdmin)}
-            className="bg-white text-blue-600 px-4 py-2 rounded hover:bg-blue-100"
-          >
-            {isAdmin ? '장바구니 페이지로' : '관리자 페이지로'}
-          </button>
-        </div>
-      </nav>
-      <main className="container mx-auto mt-6">
-        {isAdmin ? (
-          <AdminPage
-            products={products}
-            coupons={coupons}
-            onProductUpdate={handleProductUpdate}
-            onProductAdd={handleProductAdd}
-            onCouponAdd={handleCouponAdd}
-          />
-        ) : (
-          <CartPage products={products} coupons={coupons}/>
-        )}
-      </main>
-    </div>
-  );
-};
-
-export default App;
